Updated rag with file upload and file deletion.
The chat bot can accept multiple pdf files and delete either one of them or remove all of them.
The chat bot also list the available files in db.
Update the source of files to just the file name and not the file path.
Delete file in db by only retrieving the chunks of the file.